Refactored code smell with Extract Class
