+++ src/main/java/org/influxdb/querybuilder/BuiltQuery.java
@@ -29,29 +29,10 @@ public abstract class BuiltQuery extends Query implements QueryStringBuilder {
     super(null, database, requiresPost);
   }
 
-  static StringBuilder addSemicolonIfMissing(final StringBuilder stringBuilder) {
-    int length = trimLast(stringBuilder);
-    if (length == 0 || stringBuilder.charAt(length - 1) != ';') {
-      stringBuilder.append(';');
-    }
-    return stringBuilder;
-  }
-
-  static int trimLast(final StringBuilder stringBuilder) {
-    int length = stringBuilder.length();
-    while (length > 0 && stringBuilder.charAt(length - 1) <= ' ') {
-      length -= 1;
-    }
-    if (length != stringBuilder.length()) {
-      stringBuilder.setLength(length);
-    }
-    return length;
-  }
-
   @Override
   public String getCommand() {
     StringBuilder sb = buildQueryString(new StringBuilder());
-    addSemicolonIfMissing(sb);
+    Utility.addSemicolonIfMissing(sb);
     return sb.toString();
   }
